# Precompile regexes to improve performance